Minor changes to batting_leaders
I edited batting_leaders so that it pulls column headings directly from fangraphs.  This results in two separate columns having the heading 'FB%', so I renamed the second occurrence of 'FB%' to 'FB% (Pitch)', which is the same value that it had previously been assigned.

I also made a minor change to the code that creates the dataframe to make it a bit more succinct.