Minor edits to batting_leaders (#16)
* Minor fixes to pitching_leaders.py

I made two changes:
1) I fixed a bug in the hard-coded list of stats.  Previously, FB% was pulled twice: once as 49 and once as 75.  In addition, IFH% was not pulled.  Because the headings were hard-coded, this caused some headings to be mis-applied: FB% was labeled IFFB%, IFFB% was labeled HR/FB, and HR/FB was labeled IFH%.  I removed 49 from the list of stats and added 52.
2) I changed the Headings list so that it is drawn from the fangraphs output, rather than from a hard-coded list.  This allowed me to catch the bug described above.

Thanks for making this awesome package, and I hope this edit is helpful.  It's my first pull request!

* Minor changes to batting_leaders

I edited batting_leaders so that it pulls column headings directly from fangraphs.  This results in two separate columns having the heading 'FB%', so I renamed the second occurrence of 'FB%' to 'FB% (Pitch)', which is the same value that it had previously been assigned.

I also made a minor change to the code that creates the dataframe to make it a bit more succinct.