Add mlb_ID to batting_stats_range() output
This commit adds the mlb_ID of each player to the output of
batting_stats_range().  The mlb_ID allows you to cross-reference the
output with other data sources as it takes care of:
(1) common names -- e.g. multiple players named Jose Ramirez
(2) different spellings -- e.g. Enrique Hernandez in
baseball-reference.com is Kike Hernandez in the Yahoo! fantasy APIs.

The new column is appended to the very end.  The data source we pull
from already had the mlb_ID, it was buried in the href of the <a> for
the player name.  So this just uses a regex to extract that part out.