Update split_stats.py
Returning game level data as a separate dataframe if it exists. Also, upon running this against all players who started a game in September of this year, I learned that Pablo Sandoval's bbref splits stat page has an extra line in it that says "Sandoval batted solely from the left side for the last half of 2015." This breaks the height, weight, and team parsing so it has been commented out. It can be uncommented to work for other players.