Bug fix for `statcast_batter` (#256)
* Bug fix for `statcast_batter`

`statcast_batter` asserts that its parameters are not `None`, but does not reassign the parameters after the call to `sanitize_input`, leading to failure when `end_dt` is left as `None`. This commit reassigns the inputs from `sanitize_input`. `player_id` is not reassigned to avoid typing errors, since `sanitize_input` turns it into a string but `split_request` still expects an integer.

* Adds tests for statcast_batter input

Co-authored-by: Adam Kern <adam.kern@wustl.edu>